requirements.txt
openai>=1.0.0
pillow>=12.0.0
reportlab>=4.4.5
python-dotenv>=1.2.1
